updating barcode value from int to long (because it's long ey)

+++ DietTrack.SuperMarket.Core/Domain/Foods/Queries/GetFoodQuery.cs
@@ -5,8 +5,8 @@ namespace DietTrack.SuperMarket.Core.Domain.Foods.Queries
 {
     public class GetFoodQuery: IQuery<Food>
     {
-        public int Barcode { get; set; }
-        public GetFoodQuery(int barcode)
+        public long Barcode { get; set; }
+        public GetFoodQuery(long barcode)
         {
             Barcode = barcode;
         }
